# swagger.yaml
openapi: 3.0.0
info:
  title: 4TP API
  description: 4TP API
  version: 0.0.1
servers:
  - url: "https://for-the-players.herokuapp.com/"
...
paths:
  /auth/signup:
    post:
      summary: registration endpoint
      tags:
        - auth
      operationId: signup
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  description: The new user's email
                  type: string
                  format: email
                password:
                  description: The new user's password
                  type: string
                username:
                  description: The new user's password
                  type: string
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "201":
          description: "User successfully registered"
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    description: "The username of the new User"
                  email:
                    type: string
                    description: "The email of the new user"
                    format: email
                  status:
                    $ref: "#/components/schemas/UserStatuses"
                  token:
                    type: string
                    description: "The JWT of the newly created user. Only valid for the /auth/email endpoint"
  /auth/login:
    post:
      summary: login endpoint
      tags:
        - auth
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  description: The new user's email
                  type: string
                  format: email
                password:
                  description: The new user's password
                  type: string
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "200":
          description: "User successfully logged in"
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: "The JWT for the logged in user"
                  status:
                    $ref: "#/components/schemas/UserStatuses"
  /auth/email:
    get:
      summary: Send a verification email for the logged in user
      security:
        - BearerAuth: [ ]
      tags:
        - auth
      operationId: sendVerificationEmail
      responses:
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "200":
          description: "Email sent"
  /auth/password:
    post:
      summary: Sends to the user an email with a link to rest their password
      tags:
        - auth
      operationId: sendForgottenPasswordEmail
      requestBody:
        content:
          application/json:
            schema:
              required:
                - email
              type: object
              properties:
                email:
                  description: The user's email
                  type: string
                  format: email
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "200":
          description: "Email sent"
    patch:
      summary: Resets the user's password
      security:
        - BearerAuth: [ ]
      tags:
        - auth
      operationId: resetUserPassword
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - newPassword
              properties:
                newPassword:
                  type: string
                  description: The user's new password
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        200:
          description: "success"
...
  /games:
    get:
      summary: Returns the list of all games
      security:
        - BearerAuth: [ ]
      tags:
        - game
      operationId: listAllGames
      responses:
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "200":
          description: "List of games"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                      description: The Mongo ID of the game
                    name:
                      type: string
                      description: "The name of the game"
                    imgur:
                      type: string
                      description: The image for the game
    post:
      summary: Create a new game
      security:
        - BearerAuth: [ ]
      tags:
        - game
      operationId: createGame
      requestBody:
        content:
          application/json:
            schema:
              required:
                - name
              type: object
              properties:
                name:
                  description: The new game's name
                  type: string
                imgur:
                  description: >
                    The imgur link of the new game's image.
                    If this is present `imgBase64` will be ignored
                  type: string
                imgBase64:
                  description: >
                    The base64 of the new game's image.
                    If `imgur` is present this will be ignored
                  type: string
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "201":
          description: "Game successfully added"
...
  /invites/{inviteId}:
    parameters:
      - name: inviteId
        in: path
        required: true
        description: The ID of the invite
        schema:
          $ref: "#/components/schemas/MongoID"
    patch:
      security:
        - BearerAuth: [ ]
      description: Updates the status of an invite
      tags:
        - invites
      operationId: updateInvite
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                newStatus:
                  $ref: "#/components/schemas/TeamInvites"
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "404":
          description: "Invite not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "200":
          description: "Invite status updated"
    delete:
      security:
        - BearerAuth: [ ]
      description: >
        Deletes an invite
        Invite must be in status: PENDING
      tags:
        - invites
      operationId: deleteInvite
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "404":
          description: "Invite not found"
        "200":
          description: "Invite deleted"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    SchemaValidationError:
      type: array
      items:
        type: object
        properties:
          value:
            type: string
            description: "The value that was received by the server"
          msg:
            type: string
            description: "Error message"
          param:
            type: string
            description: "Name of the field that caused the error"
          location:
            type: string
            description: "Where the field was that was received by the server"
    ErrorMessageResponse:
      type: object
      properties:
        errorMessage:
          type: string
    UserStatuses:
      description: "The possible statuses for a user"
      type: string
      enum:
        - NOT_VERIFIED
        - VERIFIED
        - BANNED
...
    TeamInvites:
      description: "The possible invite statuses for team members"
      type: string
      enum:
        - PENDING
        - DECLINED
        - ACCEPTED
...
    MongoID:
      type: string
      maxLength: 24
      minLength: 24